Scheduled post
Users can schedule posts. A user can create a post and select a timestamp or delay after which the server will make the post public to all users. The server sends the user who created the post the time remaining before the post goes public every second which is displayed on their app